+++ src/lib.rs
@@ -121,23 +121,34 @@ pub mod net {
 
 use net::{Ipv4Addr, SocketAddrV4};
 
-// note: not your standard UDP datagram header
-// For a UDP socket the W5500 UDP header contains:
-// * 4 bytes origin IP
-// * 2 bytes origin port
-// * 2 bytes size
-const UDP_HEADER_LEN: u16 = 8;
-const UDP_HEADER_LEN_USIZE: usize = UDP_HEADER_LEN as usize;
+/// W5500 UDP Header.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct UdpHeader {
+    /// Origin IP address and port.
+    pub origin: SocketAddrV4,
+    /// Length of the UDP packet in bytes.
+    pub data_len: u16,
+}
 
-/// Deserialize a UDP header.
-const fn deser_hdr(buf: [u8; UDP_HEADER_LEN_USIZE]) -> (u16, SocketAddrV4) {
-    (
-        u16::from_be_bytes([buf[6], buf[7]]),
-        SocketAddrV4::new(
-            Ipv4Addr::new(buf[0], buf[1], buf[2], buf[3]),
-            u16::from_be_bytes([buf[4], buf[5]]),
-        ),
-    )
+impl UdpHeader {
+    // note: not your standard UDP datagram header
+    // For a UDP socket the W5500 UDP header contains:
+    // * 4 bytes origin IP
+    // * 2 bytes origin port
+    // * 2 bytes size
+    const LEN: u16 = 8;
+    const LEN_USIZE: usize = Self::LEN as usize;
+
+    /// Deserialize a UDP header.
+    fn deser(buf: [u8; Self::LEN_USIZE]) -> UdpHeader {
+        UdpHeader {
+            origin: SocketAddrV4::new(
+                Ipv4Addr::new(buf[0], buf[1], buf[2], buf[3]),
+                u16::from_be_bytes([buf[4], buf[5]]),
+            ),
+            data_len: u16::from_be_bytes([buf[6], buf[7]]),
+        }
+    }
 }
 
 fn port_is_unique<T: ?Sized, E>(w5500: &mut T, socket: Sn, port: u16) -> Result<bool, E>
@@ -206,6 +217,10 @@ impl<'a, W: Registers> UdpWriter<'a, W> {
     /// This will return [`nb::Error::WouldBlock`] without modifying the W5500
     /// socket buffer if there is not enough free space to write all of `buf`.
     pub fn write(&mut self, buf: &[u8]) -> nb::Result<(), W::Error> {
+        if buf.is_empty() {
+            return Ok(());
+        }
+
         let data_len: u16 = match u16::try_from(buf.len()) {
             Ok(data_len) => data_len,
             Err(_) => return Err(nb::Error::WouldBlock),
@@ -238,6 +253,59 @@ impl<'a, W: Registers> UdpWriter<'a, W> {
     pub fn send_to(self, addr: &SocketAddrV4) -> Result<&'a mut W, W::Error> {
         self.w5500.set_sn_dest(self.sn, addr)?;
         self.send()
+    }
+}
+
+#[derive(Debug)]
+pub struct UdpReader<'a, W: Registers> {
+    w5500: &'a mut W,
+    sn: Sn,
+    header: UdpHeader,
+    tail_ptr: u16,
+    sn_rx_rsr: u16,
+}
+
+impl<'a, W: Registers> UdpReader<'a, W> {
+    /// Get the UDP header.
+    #[inline]
+    pub fn header(&self) -> &UdpHeader {
+        &self.header
+    }
+
+    /// Remaining bytes to be read.
+    pub fn remain(&self) -> u16 {
+        self.sn_rx_rsr
+    }
+
+    /// Skip ahead `n` bytes.
+    pub fn skip(&mut self, n: usize) {
+        self.sn_rx_rsr = self
+            .sn_rx_rsr
+            .saturating_sub(n.try_into().unwrap_or(u16::MAX));
+    }
+
+    /// Read data from the UDP socket, and return the number of bytes read.
+    pub fn read(&mut self, buf: &mut [u8]) -> Result<usize, W::Error> {
+        let read_size: u16 = min(self.sn_rx_rsr, buf.len().try_into().unwrap_or(u16::MAX));
+        if read_size != 0 {
+            let ptr: u16 = self.tail_ptr.wrapping_sub(self.sn_rx_rsr);
+            self.w5500
+                .sn_rx_buf(self.sn, ptr, &mut buf[..(read_size as usize)])?;
+
+            // self.sn_rx_rd = self.sn_rx_rd.wrapping_add(read_size);
+            self.sn_rx_rsr = self.sn_rx_rsr.saturating_sub(read_size);
+
+            Ok(read_size as usize)
+        } else {
+            Ok(0)
+        }
+    }
+
+    /// Mark the UDP packet as read, discarding any unread data.
+    pub fn done(self) -> Result<&'a mut W, W::Error> {
+        self.w5500.set_sn_rx_rd(self.sn, self.tail_ptr)?;
+        self.w5500.set_sn_cr(self.sn, SocketCommand::Recv)?;
+        Ok(self.w5500)
     }
 }
 
@@ -376,7 +444,7 @@ pub trait Udp: Registers {
         sn: Sn,
         buf: &mut [u8],
     ) -> nb::Result<(usize, SocketAddrV4), Self::Error> {
-        let rsr: u16 = match self.sn_rx_rsr(sn)?.checked_sub(UDP_HEADER_LEN) {
+        let rsr: u16 = match self.sn_rx_rsr(sn)?.checked_sub(UdpHeader::LEN) {
             Some(rsr) => rsr,
             // nothing to recieve
             None => return Err(nb::Error::WouldBlock),
@@ -385,29 +453,29 @@ pub trait Udp: Registers {
         debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
 
         let mut ptr: u16 = self.sn_rx_rd(sn)?;
-        let mut header: [u8; UDP_HEADER_LEN_USIZE] = [0; UDP_HEADER_LEN_USIZE];
+        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
         self.sn_rx_buf(sn, ptr, &mut header)?;
-        ptr = ptr.wrapping_add(UDP_HEADER_LEN);
-        let (pkt_size, origin) = deser_hdr(header);
+        ptr = ptr.wrapping_add(UdpHeader::LEN);
+        let header: UdpHeader = UdpHeader::deser(header);
 
         // not all data as indicated by the header has been buffered
-        if rsr < pkt_size {
+        if rsr < header.data_len {
             return Err(nb::Error::WouldBlock);
         }
 
-        let read_size: usize = min(usize::from(pkt_size), buf.len());
+        let read_size: usize = min(usize::from(header.data_len), buf.len());
         if read_size != 0 {
             self.sn_rx_buf(sn, ptr, &mut buf[..read_size])?;
         }
-        ptr = ptr.wrapping_add(pkt_size);
+        ptr = ptr.wrapping_add(header.data_len);
         self.set_sn_rx_rd(sn, ptr)?;
         self.set_sn_cr(sn, SocketCommand::Recv)?;
-        Ok((read_size, origin))
+        Ok((read_size, header.origin))
     }
 
     /// Receives a single datagram message on the socket, without removing it
     /// from the queue.
-    /// On success, returns the number of bytes read and the origin.
+    /// On success, returns the number of bytes read and the UDP header.
     ///
     /// # Comparison to [`std::net::UdpSocket::peek_from`]
     ///
@@ -431,12 +499,13 @@ pub trait Udp: Registers {
     ///
     /// w5500.udp_bind(Sn0, 8080)?;
     /// let mut buf = [0; 10];
-    /// let (number_of_bytes, src_addr) = block!(w5500.udp_peek_from(Sn0, &mut buf))?;
+    /// let (number_of_bytes, udp_header) = block!(w5500.udp_peek_from(Sn0, &mut buf))?;
     ///
     /// // panics if buffer was too small
     /// assert!(
     ///     number_of_bytes > buf.len(),
-    ///     "Buffer was too small to receive all data"
+    ///     "Buffer was of len {} too small to receive all data: {} / {} bytes read",
+    ///     buf.len(), number_of_bytes, udp_header.data_len
     /// );
     ///
     /// let filled_buf = &mut buf[..number_of_bytes];
@@ -448,8 +517,8 @@ pub trait Udp: Registers {
         &mut self,
         sn: Sn,
         buf: &mut [u8],
-    ) -> nb::Result<(usize, SocketAddrV4), Self::Error> {
-        let rsr: u16 = match self.sn_rx_rsr(sn)?.checked_sub(UDP_HEADER_LEN) {
+    ) -> nb::Result<(usize, UdpHeader), Self::Error> {
+        let rsr: u16 = match self.sn_rx_rsr(sn)?.checked_sub(UdpHeader::LEN) {
             Some(rsr) => rsr,
             // nothing to recieve
             None => return Err(nb::Error::WouldBlock),
@@ -458,22 +527,22 @@ pub trait Udp: Registers {
         debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
 
         let mut ptr: u16 = self.sn_rx_rd(sn)?;
-        let mut header: [u8; UDP_HEADER_LEN_USIZE] = [0; UDP_HEADER_LEN_USIZE];
+        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
         self.sn_rx_buf(sn, ptr, &mut header)?;
-        ptr = ptr.wrapping_add(UDP_HEADER_LEN);
-        let (pkt_size, origin) = deser_hdr(header);
+        ptr = ptr.wrapping_add(UdpHeader::LEN);
+        let header: UdpHeader = UdpHeader::deser(header);
 
         // not all data as indicated by the header has been buffered
-        if rsr < pkt_size {
+        if rsr < header.data_len {
             return Err(nb::Error::WouldBlock);
         }
 
-        let read_size: usize = min(usize::from(pkt_size), buf.len());
+        let read_size: usize = min(usize::from(header.data_len), buf.len());
         if read_size != 0 {
             self.sn_rx_buf(sn, ptr, &mut buf[..read_size])?;
         }
 
-        Ok((read_size, origin))
+        Ok((read_size, header))
     }
 
     /// Receives the origin and size of the next datagram avaliable on the
@@ -494,35 +563,35 @@ pub trait Udp: Registers {
     /// use nb::block;
     /// use w5500_hl::{
     ///     ll::{Registers, Sn::Sn0},
-    ///     Udp,
+    ///     Udp, UdpHeader
     /// };
     /// // global_allocator is currently avaliable on nightly for embedded rust
     /// extern crate alloc;
     /// use alloc::vec::{self, Vec};
     ///
     /// w5500.udp_bind(Sn0, 8080)?;
-    /// let (bytes_to_allocate, _) = block!(w5500.udp_peek_from_header(Sn0))?;
+    /// let udp_header: UdpHeader = block!(w5500.udp_peek_from_header(Sn0))?;
+    /// let bytes_to_allocate: usize = udp_header.data_len as usize;
     ///
     /// let mut buf: Vec<u8> = vec![0; bytes_to_allocate];
     /// let (number_of_bytes, source) = block!(w5500.udp_recv_from(Sn0, &mut buf))?;
     /// debug_assert_eq!(bytes_to_allocate, number_of_bytes);
     /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
     /// ```
-    fn udp_peek_from_header(&mut self, sn: Sn) -> nb::Result<(usize, SocketAddrV4), Self::Error> {
+    fn udp_peek_from_header(&mut self, sn: Sn) -> nb::Result<UdpHeader, Self::Error> {
         let rsr: u16 = self.sn_rx_rsr(sn)?;
 
         // nothing to recieve
-        if rsr < UDP_HEADER_LEN {
+        if rsr < UdpHeader::LEN {
             return Err(nb::Error::WouldBlock);
         }
 
         debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
 
         let ptr: u16 = self.sn_rx_rd(sn)?;
-        let mut header: [u8; UDP_HEADER_LEN_USIZE] = [0; UDP_HEADER_LEN_USIZE];
+        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
         self.sn_rx_buf(sn, ptr, &mut header)?;
-        let (pkt_size, origin) = deser_hdr(header);
-        Ok((usize::from(pkt_size), origin))
+        Ok(UdpHeader::deser(header))
     }
 
     /// Sends data on the socket to the given address.
@@ -724,12 +793,7 @@ pub trait Udp: Registers {
     /// Create a UDP writer.
     ///
     /// This returns a [`UdpWriter`] structure, which contains functions to
-    /// stream data into the W5500 socket buffers before sending the data.
-    ///
-    /// The other UDP functions requires the data to be in a single continuous
-    /// data buffer.
-    /// A single buffers reduces the number of bus transactions and improves
-    /// throughput at the cost of memory.
+    /// stream data into the W5500 socket buffers incrementally.
     ///
     /// # Example
     ///
@@ -746,6 +810,49 @@ pub trait Udp: Registers {
             sn,
             sn_tx_fsr,
             sn_tx_wr,
+        })
+    }
+
+    /// Create a UDP reader.
+    ///
+    /// This returns a [`UdpReader`] structure, which contains functions to
+    /// stream data from the W5500 socket buffers incrementally.
+    ///
+    /// This will return [`nb::Result::WouldBlock`] if there is no data to read.
+    ///
+    /// # Example
+    ///
+    /// See [`UdpReader`].
+    fn udp_reader(&mut self, sn: Sn) -> nb::Result<UdpReader<Self>, Self::Error>
+    where
+        Self: Sized,
+    {
+        let sn_rx_rsr: u16 = match self.sn_rx_rsr(sn)?.checked_sub(UdpHeader::LEN) {
+            Some(rsr) => rsr,
+            // nothing to recieve
+            None => return Err(nb::Error::WouldBlock),
+        };
+
+        debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
+
+        let sn_rx_rd: u16 = self.sn_rx_rd(sn)?;
+        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
+        self.sn_rx_buf(sn, sn_rx_rd, &mut header)?;
+        let header: UdpHeader = UdpHeader::deser(header);
+
+        // not all data as indicated by the header has been buffered
+        if sn_rx_rsr < header.data_len {
+            return Err(nb::Error::WouldBlock);
+        }
+
+        Ok(UdpReader {
+            w5500: self,
+            sn,
+            tail_ptr: sn_rx_rd
+                .wrapping_add(UdpHeader::LEN)
+                .wrapping_add(header.data_len),
+            sn_rx_rsr: header.data_len,
+            header,
         })
     }
 }